+++ .gitignore
@@ -3,4 +3,4 @@ node_modules
 prisma/dev.db
 .vscode
 prisma/dev.db-journal
-public/*
+public
